requirements.txt
pydub==0.25.1
joblib==1.0.1
aiofiles
lightgbm==3.2.1
python-multipart
absl-py==0.13.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
cachetools==4.2.2
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0
decorator==5.0.9
fsspec==2021.8.1
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.39.0
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.8.1
joblib==1.0.1
librosa==0.8.1
llvmlite==0.36.0
markdown==3.3.4
multidict==5.1.0
numba==0.53.1
oauthlib==3.1.1
packaging==21.0
pandas==1.1.5
pooch==1.5.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.2.8
pytz==2021.1
pyyaml==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
soundfile==0.10.3.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
torchlibrosa==0.0.9
torchmetrics==0.5.1
tqdm==4.62.2
urllib3==1.26.6
werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
uvicorn==0.15.0
pydantic==1.8.2
fastapi==0.68.1